Clean: unused files and comments
